04/18/24
implemented fully functioning firework array.

rectified a bug in which only one firework would be drawn - and another in which firework speed variables would only be one value across the entire array.

rectified an error in which gravity was not assigned - and another in which pauseUpdate(); was not ran - thus, the firework array would continue to draw when it was not supposed to.

in summary, do not EVER touch the firework class again if you don't want grey hair...