04/21/24
rectified redundant one/two player buttons -- now, they only appear when a new game is created, and are much more eye-catching for the player to see.

implemented println to keep track of one/two player selection in the console.

rectified a known bug in which the player could select one or two players while a game was in session by the implementation of an error catch

implemented a known bug in which the player could move the ball while the game was paused by the implementation of an error catch.