04/23/24 V2
tried and (semi-)failed to implement multiple balls .... idk if i want that feature. i dont think it's very fun and it just seems to cause issues, however, i commented out the code for this feature in case I change my mind.

rectified (?) a known bug in which textSize was hard-coded and would not display on some display sizes.

TRY??